requirements_test.txt
pylint==2.5.2
pylint-flask-sqlalchemy==0.2.0
pytest==5.4.2
